backend/requirements.txt
fastapi>=0.115.0
uvicorn>=0.31.0
pdfplumber>=0.11.4
spacy>=3.7.6
sentence-transformers>=3.1.1
faiss-cpu>=1.8.0
requests>=2.32.3
pydantic>=2.9.2
sqlmodel==0.0.22 
psycopg2-binary==2.9.10
pg8000==1.30.1
alembic==1.13.3 
python-jose[cryptography]==3.3.0 
python-dotenv==1.0.1
slowapi>=0.1.9 
passlib[bcrypt]==1.7.4
asyncpg==0.30.0